+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 <!--next-version-placeholder-->
 
+## v5.1.1 (2026-09-16)
+
+### Fix
+* **query:** warn on no-op builder calls
+
+## v5.1.0 (2026-09-08)
+
+### Feature
+* **query:** add async iteration and windowed query results (#13)
+
 ## v5.0.6 (2026-09-01)
 
 ### Fix